Add CI for pull requests and when code lands on main
When code lands on main, the images are published with the `tip` tag.
When a pull request is submitted smoke tests are run with the func client
API. A function is created for each of the builtin templates for the
buildpack runtimes (http/events for go/typescript), and then built.

Needs: https://github.com/knative-sandbox/kn-plugin-func/pull/495

Signed-off-by: Lance Ball <lball@redhat.com>