feat: adds a new buildpack to detect func.yaml
This commit adds a new buildpack which has a singular purpose of checking
for a func.yaml file in the project directory and failing if one does not
exist.

* `ghcr.io/boson-project/boson-function-buildpack`

To use it, add the following to your buildpack's build plan.

```
[[requires]]
  name = "boson"
```

This commit introduces that dependency on the existing Go buildpack.

The `detect` script could be improved to check for additional metadata in
the func.yaml file if we feel that a check for simple existence isn't enough.

A test has been added that introduces the `pack` dependency in the build
to test this buildpack. An additional test needs to be added to test for
the failure case, when the project directory does not include a func.yaml.

Signed-off-by: Lance Ball <lball@redhat.com>